+++ docs/ReleaseNotes.md
@@ -1,10 +1,14 @@
 # `WeihanLi.Redis` ReleaseNotes
 
-## 1.5.4
+## [WeihanLi.Redis 1.5.6](https://www.nuget.org/packages/WeihanLi.Redis/1.5.6)
+
+- Add `RedisEventBus` and `RedisEventStore`
+
+## [WeihanLi.Redis 1.5.4](https://www.nuget.org/packages/WeihanLi.Redis/1.5.4)
 
 - Add Multi Get for HashClient
 
-## 1.5.3
+## [WeihanLi.Redis 1.5.3](https://www.nuget.org/packages/WeihanLi.Redis/1.5.3)
 
 - Add HashCounterClient
 - Add support for custom `IDataSerializer` and custom `IDataCompressor`